--- src/utils/initialProfiles.js
@@ -1,28 +1,0 @@
-const initialProfiles = [
-  {
-    id: 'A',
-    profile: 'readerViewA',
-    name: 'Light',
-    backgroundColor: '#FFFDD0',
-    color: '#202020',
-    linkColor: '#3101ee',
-  },
-  {
-    id: 'B',
-    profile: 'readerViewB',
-    name: 'Dark',
-    backgroundColor: '#202020',
-    color: '#eeeeee',
-    linkColor: '#8ab4f8',
-  },
-  {
-    id: 'C',
-    profile: 'readerViewC',
-    name: 'Sepia',
-    backgroundColor: '#ffeedd',
-    color: '#332211',
-    linkColor: '#3101ee',
-  },
-]
-
-export default initialProfiles
