--- src/utils/initialSettings.js
@@ -1,12 +1,0 @@
-const initialSettings = {
-  textAlign: 'start',
-  fontFamily: 'sans-serif',
-  fontSize: 1,
-  lineHeight: 1.5,
-  wordSpacing: 0.2,
-  letterSpacing: 0.08,
-  maxWidth: 40,
-  linkColor: '#3101ee',
-  blockImages: false,
-}
-export default initialSettings
